FAIL: Attempt to impl throughput in display…
…which can't work because it's actually never mutable due to the way
drawing work: it operates on a snapshot, a copy, that is not written
back.

And even if it was, the type system statically concludes sync is needed
as well for this to work.

Long story short: No state changes are ever allowed with a system like
this, and throughput needs to maintain just that.

Throughput must be implemented in each renderer.