feat!: Allow rendererers to respond to dropped progress roots
Previously it needed extra effort to communicate that a the computation
was done and the renderer should stop rendering progress.

Now they only obtain a weak progress instance so it can drop if the
computation is done, terminating it naturally and in time.

Note that in case of the TUI, it would still be needed to respond
to the GUI having shut down due to user request.