feat: Improve render-log performance greatly.
Previously it would check the current time each time somebody
wants to log on any logger, greatly reducing performance as
it would block on the mutex rust-std uses internally.

Now we use a single thread to provide information about whether or not
we may log, whose lifetime is bound to all of the log instances it
governs.