feat: Introduce the object-safe `RawProgress` trait.
It's automatically implemented for `Progress` and allows for more flexible use
of progress particularly in leaf nodes. This is useful if a function needs to take
multiple types of progress as it is called from different places in the same function.

Without dyn-traits, it's not possible to make such call.