Bump `bytesize` from major version 1 to 2
This upgrades the `bytesize` dependency in `Cargo.toml` from
version 1.0.1 (which was usually selecting 1.3.3) to 2.0.1 (which
is the latest version).

There were some nontrivial API changes from major version 1 to 2.
Accordingly, this adapts `Bytes::format_bytes()` to the API change.

The old `bytesize::to_string()` function was confusing in its
second parameter `si_prefix: bool`, which is why it was removed.
But it looks like a value of `false`, as we were passing, actually
caused decimal SI units (rather than binary IEC units) to be used.
It's not obvious which units `prodash` intended to use `bytesize`
to convert to and display in. But this seems to be a minimal change
to adapt to the new major version.

In spite of the name `si_prefix` for the old parameter, experiments
show that setting it to `true` caused values to be converted to and
presented in binary IEC units.

Although this attempts not to change which actual units are used,
it does produce observable differences for some of them, in how
they are presented. In particular, a decimal SI kilobyte is least
ambiguously abbreviated "kB" (because "k" is an SI unit symbol
prefix for "kilo" in its meaning of 1000), but it was previously
written as "KB". It is now expected to be writen as "kB". Tests
catch this distinction, and are updated here accordingly to assert
that the generally preferable "kB" symbol for decimal SI kilobyte
is used.

See also:
https://github.com/GitoxideLabs/gitoxide/issues/1947#issuecomment-2795974907